Working on SAX hashing

+++ utils.py
@@ -1,9 +1,9 @@
 import numpy as np
 import time
 cut_points = {
-    2 : np.array([np.NINF, 0]),
-    3 : np.array([np.NINF,-0.43,0.43]),
-    4 : np.array([np.NINF,-0.67,0,0.67])
+    2 : np.array([np.NINF, 0],dtype=np.float32),
+    3 : np.array([np.NINF,-0.43,0.43],dtype=np.float32),
+    4 : np.array([np.NINF,-0.67,0,0.67],dtype=np.float32)
 }
 
 # Simple decorator to time each part of program execution
